add: コンポーネントに分割
