+++ NearU/Model/SkillTagRepository.swift
@@ -7,8 +7,8 @@
 
 import Foundation
 
-struct TechTagRepository {
-    static let techTags: [String: [String]] = [
+struct SkillTagRepository {
+    static let skillTags: [String: [String]] = [
         "言語": ["Swift", "Kotlin", "JavaScript", "Python", "Go", "Java", "Ruby", "PHP", "TypeScript", "C", "C++", "C#", "HTML", "CSS", "Rust", "Dart", "Elixir"],
         "フレームワーク": ["React", "Next.js", "Vue", "Nuxt.js", "Angular", "Node.js", "Django", "Flask", "Laravel", "CakePHP", "Flutter", "Rails", "Remix", "Tailwind CSS", "Spring"],
         "データベース": ["MySQL", "PostgreSQL", "MongoDB", "Redis", "MariaDB", "DynamoDB"],
