--- NearU/View/Profile/View/ProfileCompornents/TechTagPickerView.swift
@@ -1,18 +1,0 @@
-//
-//  TechTagPickerView.swift
-//  NearU
-//
-//  Created by 高橋和 on 2025/03/23.
-//
-
-import SwiftUI
-
-struct TechTagPickerView: View {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-#Preview {
-    TechTagPickerView()
-}
